Fix directory permissions
This commit makes this package to be in line with the package
'icingaweb2' when it comes to directory permissions under
'/etc/icingaweb2'.
This way the Icinga Director can actually deploy its own 'config.ini'
when choosing a DB Resource in the Icinga Web 2.

It is assumed that the 'icingaweb2' package correctly creates a
'icingaweb2' group.